+++ tests/examples/rust/src/lib.rs
@@ -1,0 +1,14 @@
+#[path = "../../../../examples/authors/rust/native/lib.rs"]
+pub mod authors;
+
+#[path = "../../../../examples/batch/rust/native/lib.rs"]
+pub mod batch;
+
+#[path = "../../../../examples/booktest/rust/native/lib.rs"]
+pub mod booktest;
+
+#[path = "../../../../examples/jets/rust/native/lib.rs"]
+pub mod jets;
+
+#[path = "../../../../examples/ondeck/rust/native/lib.rs"]
+pub mod ondeck;